fix scroll bar
